requirements.txt
streamlit==1.24.0
PyPDF2==3.0.1
deep-translator==1.11.4
google-generativeai==0.3.1
reportlab==3.6.12
python-dotenv==1.0.0